chore: Move verify endpoint to the endpoint section of app
